Functional updates and bug fixes
Added the ability to import a certificate directly into the application without additional interaction with the system (useful for debugging applications on Android TV, where the functionality is severely limited).

There should be fewer bugs now. The build script is updated with possible exceptions and solutions (of course, only those that I encountered at the moment). Also, a zipalign step has been added to the build process.

In addition, comments have been added to the scripts for the future me and the users of the tool to understand the processes.